Refatorada aplicação separando produtos e vendas
close #40
