+++ produtos/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ProdutosConfig(AppConfig):
+    name = 'produtos'
