+++ produtos/models.py
@@ -1,0 +1,9 @@
+from django.db import models
+
+# Create your models here.
+class Produto(models.Model):
+    descricao = models.CharField(max_length=100)
+    preco = models.DecimalField(max_digits=5, decimal_places=2)
+
+    def __str__(self):
+        return self.descricao
